Components: Extract a reusable Post Format Selector
