+++ editor/post-format/style.scss
@@ -1,12 +1,14 @@
 .editor-post-format {
 	flex-direction: column;
 	align-items: stretch;
+	width: 100%;
 }
 
 .editor-post-format__content {
 	display: inline-flex;
 	justify-content: space-between;
 	align-items: center;
+	width: 100%;
 }
 
 .editor-post-format__suggestion {
